Former-commit-id: 2f85d2877e284409fa8637380290eae3d12bfc60

+++ src/main/java/cz/cvut/fel/aic/amodsim/entity/vehicle/OnDemandVehicle.java
@@ -359,8 +359,8 @@ public class OnDemandVehicle extends Agent implements EventHandler, PlanningAgen
 	}
 
 	@Override
-	public int getVelocity() {
-		return config.vehicleSpeedInMeters;
+	public double getVelocity() {
+		return (double) config.vehicleSpeedInMeters;
 	}
 	
 	public int getCapacity(){
